[jira] [HBASE-5274] Filter out expired scanners on compaction as well
Summary: This is a followup for D1017 to make it similar to D909 (89-fb). The
fix for 89-fb used the TTL-based scanner filtering logic on both normal scanners
and compactions, while the trunk fix D1017 did not. This is just the delta
between the two diffs that brings filtering expired store files on compaction to
trunk.

Test Plan: Unit tests

Reviewers: Liyin, JIRA, lhofhansl, Kannan

Reviewed By: Liyin

CC: Liyin, tedyu, Kannan, mbautin, lhofhansl

Differential Revision: https://reviews.facebook.net/D1473

git-svn-id: https://svn.apache.org/repos/asf/hbase/trunk@1236483 13f79535-47bb-0310-9956-ffa450edef68