[jira] [HBASE-5744] Thrift server metrics should be long instead of int
Summary:
As we measure our Thrift call latencies in nanoseconds, we need to make
latencies long instead of int everywhere. There is a bug where we truncate a
nanosecond latency to int, which is a problem with RPCs that take more than
2.147483647 seconds to process.

This is a trunk diff. The 89-fb version of this diff is at D2679.

Test Plan:
TestThriftServer is updated to test for the failure case (an RPC is artificially
made to take 3 seconds). The new test case fails without the fix.

Re-run all unit tests.

Reviewers: stack, sc, Kannan, JIRA

Reviewed By: sc

Differential Revision: https://reviews.facebook.net/D2685



git-svn-id: https://svn.apache.org/repos/asf/hbase/trunk@1311167 13f79535-47bb-0310-9956-ffa450edef68