HBASE-9893 Incorrect assert condition in OrderedBytes decoding
Correct an invalid assumption in remaining assertion code around
OrderedBytes#decodeVarBlob. When an encoded value contains a 1-bit in its LSB
position and the length of the encoded byte array is divisible by 7, the value
remaining in variable t will be 0x80, resulting in the failed assertion coming
out of the decoding loop. This patch preserves the assertion for the general
case by resetting 't' at the conclusion of the 7-byte cycle.


git-svn-id: https://svn.apache.org/repos/asf/hbase/trunk@1543699 13f79535-47bb-0310-9956-ffa450edef68